sys: Attempt at restructuring the source tree
The goal here is to make the src and sys builds more independent of
their location by changing the directory structure of sys to match that
of FreeBSD.

This will allow symlinks from src/include to src/sys and in turn allow
the src build to use its own headers instead of those in the system
environment.

As far as patches are concerned this is a breaking change. Advice is to
put patches on the 'historic' branch and rebase this one if needed.